requirements.txt
Flask==0.11.1
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.15
Werkzeug==0.11.11
argparse==1.2.1
click==6.6
itsdangerous==0.24
pycrypto==2.6.1
pysqlite==2.8.3
wsgiref==0.1.2
flake8==3.0.4
